code/requirements.txt
commonroad_drivability_checker==2021.4
commonroad_io==2021.4
commonroad_route_planner==1.0.0
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.5
Pillow==9.2.0
scipy==1.7.3
setuptools==58.0.4
Shapely==1.8.0
